Cleanup
Update the readme, finetune the example
