+++ src/openapi.js
@@ -8,10 +8,8 @@ import { readFile } from 'node:fs/promises'
  * @param {string=} params.base
  * @returns {Promise<{ openAPISpecification: object; }>}
  */
-const openAPI = async ({ file, base = import.meta.url }) => {
+export const openAPI = async ({ file, base = import.meta.url }) => {
   const fileUrl = new URL(file, base)
   const openAPISpecification = JSON.parse(await readFile(fileUrl, 'utf8'))
   return { openAPISpecification }
 }
-
-export { openAPI }
