+++ src/params.js
@@ -7,16 +7,16 @@ import { types } from './types.js'
  * @param {object} params.spec
  * @returns {object}
  */
-const parseParams = ({ query, spec }) => spec.map(parameter => {
-  const { name, schema } = parameter
-  const { type } = schema
-  const Type = types[type]
-  const paramName = query[name]
-  const value = new Type(paramName).valueOf()
-  return { name, value }
-}).reduce((acc, { name, value }) => {
-  acc[name] = value
-  return acc
-}, {})
-
-export { parseParams }
+export const parseParams = ({ query, spec }) =>
+  spec.map(parameter => {
+    const { name, schema } = parameter
+    const { type } = schema
+    const Type = types[type]
+    const paramName = query[name]
+    const value = new Type(paramName).valueOf()
+    return { name, value }
+  })
+    .reduce((acc, { name, value }) => {
+      acc[name] = value
+      return acc
+    }, {})
