+++ src/router.js
@@ -1,60 +1,10 @@
 import { OpenAPIBackend } from 'openapi-backend'
-import getStatusByError from './error-status.js'
-import { parseParams } from './params.js'
-
-const makeExpressCallback = ({
-  controller,
-  specification
-}) =>
-/**
- * Handle controller
- * @async
- * @param {object} context
- * @param {object} request
- * @param {object} response
- * @returns {Promise<any>}
- */
-  async (context, request, response) => {
-    const parameters = parseParams({
-      query: context.request.query,
-      spec: context.operation.parameters
-    })
-    const url = `${request.protocol}://${request.get('Host')}${request.originalUrl}`
-
-    try {
-      return controller({
-        context,
-        request,
-        response,
-        parameters,
-        specification,
-        url
-      })
-    } catch (error) {
-      const errorCodeStatus = getStatusByError(error)
-      response.status(errorCodeStatus)
-      return {
-        status: errorCodeStatus,
-        timestamp: new Date(),
-        message: error.message
-      }
-    }
-  }
-
-const operations = ['get', 'put', 'patch', 'post', 'delete']
-
-/**
- * Get all operation ID's from the specification.
- * @param {object} params
- * @param {object} params.specification
- * @returns {string[]}
- */
-const operationIds = ({ specification }) => Object.values(specification.paths)
-  .map((path) => Object.entries(path)
-    .map(([operation, data]) => operations.includes(operation)
-      ? data.operationId
-      : null))
-  .flat()
+import { makeExpressCallback } from './express-callback.js'
+import { operationIds } from './operation-ids.js'
+import { notFound } from './handlers/not-found.js'
+import { requestValidation } from './handlers/request-validation.js'
+import { responseValidation } from './handlers/response-validation.js'
+import { unauthorized } from './handlers/unauthorized.js'
 
 /**
  * Setup the router
@@ -64,60 +14,16 @@ const operationIds = ({ specification }) => Object.values(specification.paths)
  * @param {object} params.controllers
  * @returns {{ api, openAPISpecification: object }}
  */
-const setupRouter = ({ env, openAPISpecification, controllers }) => {
+export const setupRouter = ({ env, openAPISpecification, controllers }) => {
   const secret = env.SECRET
 
   const api = new OpenAPIBackend({ definition: openAPISpecification })
 
   api.register({
-    unauthorizedHandler: async (context, request, response) => {
-      response.status(401)
-      return {
-        status: 401,
-        timestamp: new Date(),
-        message: 'Unauthorized'
-      }
-    },
-    // Request validation
-    validationFail: (context, request, response) => {
-      response.status(400)
-      return {
-        errors: context.validation.errors,
-        status: 400,
-        timestamp: new Date(),
-        message: 'Bad Request'
-      }
-    },
-    notFound: (_context, request, response) => {
-      response.status(404)
-      return {
-        status: 404,
-        timestamp: new Date(),
-        message: 'Not found'
-      }
-    },
-    // Response validation
-    postResponseHandler: (context, request, response) => {
-      const responseDoesntNeedValidation = response.statusCode >= 400
-      if (responseDoesntNeedValidation) {
-        return response.json(context.response)
-      }
-
-      const valid = context.api.validateResponse(
-        context.response,
-        context.operation
-      )
-      if (valid?.errors) {
-        return response.status(502).json({
-          errors: valid.errors,
-          status: 502,
-          timestamp: new Date(),
-          message: 'Bad response'
-        })
-      }
-
-      return response.json(context.response)
-    }
+    unauthorizedHandler: unauthorized,
+    validationFail: requestValidation,
+    notFound,
+    postResponseHandler: responseValidation
   })
 
   operationIds({ specification: openAPISpecification }).forEach((operationId) => {
@@ -137,5 +43,3 @@ const setupRouter = ({ env, openAPISpecification, controllers }) => {
 
   return { api, openAPISpecification }
 }
-
-export { setupRouter }
